Adds most of the functionality for palette picker
Updates styles for all the menu and project menu sections. Added a delete icon for palettes. Adds functionality to the front end to both add and delete palettes and projects. Updates the knex migration file to allow for onDelete cascade. Adds methods to jquery that populate the project drop down and the projects palettes that are in the database. Adds functions to the script.js that add and remove projects and palettes from the database while also updating the dom on the fly. Whew...